virtio: remove LegacyVirtioDevice trait and LegacyWrapper
With virtio_p9 modernized (Phase 1) and virtio_serial deleted (Phase 2),
the only remaining LegacyVirtioDevice user was TestDevice in tests.rs.

Convert TestDevice to implement VirtioDevice directly with enable/disable,
following the same pattern used by real devices (driver-owned workers,
exit_event for shutdown). This removes all LegacyVirtioDevice usage.

Remove from common.rs:
- LegacyVirtioDevice trait
- LegacyWrapper struct and all impls (VirtioDevice, Drop)
- VirtioState and VirtioRunningState enums (only used by LegacyWrapper)
- wait_for_outstanding_descriptors (dead after LegacyWrapper removal)
- await_outstanding_descriptors on VirtioQueueUsedHandler (dead)
- Unused imports (Spawn, VmTaskDriver, VmTaskDriverSource)

Also fix pre-existing unused tracing_helpers dep in openvmm_core.

All 14 virtio tests pass. This is Phase 3 of the legacy virtio removal.
